Move newWindow() to RemoteTargetLocator, add option to specify window type
