+++ lib/Remote/RemoteTargetLocator.php
@@ -2,6 +2,7 @@
 
 namespace Facebook\WebDriver\Remote;
 
+use Facebook\WebDriver\Exception\UnsupportedOperationException;
 use Facebook\WebDriver\WebDriver;
 use Facebook\WebDriver\WebDriverAlert;
 use Facebook\WebDriver\WebDriverElement;
@@ -112,6 +113,33 @@ class RemoteTargetLocator implements WebDriverTargetLocator
         }
 
         $this->executor->execute(DriverCommand::SWITCH_TO_WINDOW, $params);
+
+        return $this->driver;
+    }
+
+    /**
+     * Creates a new browser window and switches the focus for future commands of this driver to the new window.
+     *
+     * @see https://w3c.github.io/webdriver/#new-window
+     * @param string $windowType The type of a new browser window that should be created. One of [tab, window].
+     * The created window is not guaranteed to be of the requested type; if the driver does not support the requested
+     * type, a new browser window will be created of whatever type the driver does support.
+     * @throws UnsupportedOperationException
+     * @return RemoteWebDriver This driver focused on the given window
+     */
+    public function newWindow($windowType = self::WINDOW_TYPE_TAB)
+    {
+        if ($windowType !== self::WINDOW_TYPE_TAB && $windowType !== self::WINDOW_TYPE_WINDOW) {
+            throw new \InvalidArgumentException('Window type must by either "tab" or "window"');
+        }
+
+        if (!$this->isW3cCompliant) {
+            throw new UnsupportedOperationException('New window is only supported in W3C mode');
+        }
+
+        $response = $this->executor->execute(DriverCommand::NEW_WINDOW, ['type' => $windowType]);
+
+        $this->window($response['handle']);
 
         return $this->driver;
     }
